Refactor: Refactor client side folder structure
