+++ fe/src/components/modal/Delete.tsx
@@ -1,4 +1,4 @@
-import { Instance } from "../config/Instance";
+import { Instance } from "../../utils/Instance";
 
 type TProps = {
   setIsDeleteModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
@@ -13,11 +13,8 @@ export default function Delete({
   console.log("delete url", deleteUrl);
   const handleDelete = async () => {
     try {
-
       // const res = await Instance.delete(`/v1/service/${selectDeleteId}`);
       await Instance.delete(deleteUrl);
-
-
 
       onDeleteSuccess();
     } catch (error) {
