+++ fe/src/interfaces/types/allowance.types.ts
@@ -1,0 +1,13 @@
+export type TAllowance = {
+  allowance_CD: string;
+  allowance_description: string;
+  allowance_nepali_desc: string;
+  allowance_facility: string;
+  allowance_taxable: string;
+  allowance_facility_percent: string;
+  allowance_cit_flag: string;
+  allowance_type: string;
+  salary_allowance_flag: string;
+  allowance_acc_cd: string;
+  allowance_disabled: string;
+};
