+++ fe/src/interfaces/types/login.types.ts
@@ -1,0 +1,4 @@
+export type TLogin = {
+  username: string;
+  hashedPassword: string;
+};
