+++ fe/src/interfaces/types/serviceEvent.types.ts
@@ -1,0 +1,12 @@
+export type TServiceEvent = {
+  SERVICE_EVENT_CD: string;
+  SERVICE_EVENT_DESC: string;
+  SERVICE_EVENT_DESC_NEP: string | null;
+  DISABLED: string;
+  ENTERED_BY: string;
+  ENTERED_DT: string;
+  SERVICE_EVENT_TYPE: string;
+  IS_AUTO_SALARY_ADJUST: string;
+  LAST_UPDATED_BY: string | null;
+  LAST_UPDATED_ON: string | null;
+};
