+++ fe/src/interfaces/types/sidebar.types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+export type TsidebarItems = {
+  title: string;
+  icon: () => ReactNode;
+  path: string;
+};
